test(select): Refactor visual testing stories to use CSF
